Add RefreshOIDCAccessToken middleware
The OP can provide a refresh_token to the client on authentication. This
can later be used to get a new access_token. Typically refresh_tokens
have a longer TTL than access_tokens and represent the total allowed
session length. As a bonus, the refresh happens in the background and
does not require taking the user to a new location (which also makes it
more compatible with e.g., XHR).

If any error occurs during refresh, the middleware aborts, but does not
perform any cleanup on the session.